Split resource.rb into two files
