+++ lib/heimdallr-resource.rb
@@ -1,0 +1,2 @@
+require 'heimdallr/resource'
+require 'heimdallr/resource_implementation'
